+++ python/deptry/violations/base.py
@@ -23,12 +23,13 @@ class ViolationsFinder(ABC):
         dependencies: A list of Dependency objects representing the project's dependencies.
         ignored_modules: A tuple of module names to ignore when scanning for issues. Defaults to an
             empty tuple.
-
+        standard_library_modules: A set of modules that are part of the standard library
     """
 
     violation: ClassVar[type[Violation]]
     imported_modules_with_locations: list[ModuleLocations]
     dependencies: list[Dependency]
+    standard_library_modules: frozenset[str]
     ignored_modules: tuple[str, ...] = ()
 
     @abstractmethod