+++ python/deptry/violations/dep001_missing/finder.py
@@ -27,6 +27,9 @@ class DEP001MissingDependenciesFinder(ViolationsFinder):
         for module_with_locations in self.imported_modules_with_locations:
             module = module_with_locations.module
 
+            if module.standard_library:
+                continue
+
             logging.debug("Scanning module %s...", module.name)
 
             if self._is_missing(module):