+++ tests/data/project_using_namespace/foo/database/bar.py
@@ -2,7 +2,6 @@ from os import chdir, walk
 from pathlib import Path
 
 import black
-import click
 import white as w
-from urllib3 import contrib
-import asyncio
+
+from foo import api