+++ tests/data/some_imports.py
@@ -2,6 +2,10 @@ import typing
 from os import chdir, walk
 from pathlib import Path
 from typing import List, TYPE_CHECKING
+from importlib import import_module
+from importlib import import_module as im
+import importlib
+import importlib as il
 
 import numpy as np
 import pandas
@@ -40,3 +44,9 @@ def func():
 class MyClass:
     def __init__(self):
         import module_in_class
+
+import_module("patito")
+importlib.import_module("polars")
+im("uvicorn")
+import_module("http.server")
+il.import_module("xml.etree.ElementTree")